Restructure and display player controls
Restructured code for maintainability (split into modular sections);
Added player controls to start screen;
