+++ src/pypong.py
@@ -1,0 +1,6 @@
+from game.game import Game
+
+if __name__ == "__main__":
+    game = Game()
+    game.run()
+
